Redesign configuration API
