+++ packages/cli/src/@internal/commands/configCommand.ts
@@ -1,7 +1,7 @@
 import convict from 'convict'
 import commander from 'commander'
 import { formatDocs, getDocs } from '../config/docs'
-import { validateConfig, loadConfigIfExists } from '../config'
+import { validateConfig, loadConfigIfExists } from '../config/utils'
 
 export const configCommand = (config: convict.Config<any>) => (cli: commander.Command) => cli
     .command('config <command>')
