+++ packages/cli/src/configuration.ts
@@ -1,1 +1,2 @@
-export { createSchema, Schema, REQUIRED } from './@internal/config/index'
+export * from './@internal/config/operators'
+export * from './@internal/config/types'
