+++ packages/cli/src/index.ts
@@ -3,5 +3,4 @@ import { polyfill } from './@internal/polyfills'
 polyfill(global)
 
 export { EtlPipe } from './@internal/pipe'
-export { EtlxBuilder, etlx } from './@internal/builder/etlx'
-export { ConfigurationBuilder } from './@internal/builder/configuration'
+export { etlx } from './@internal/builder/etlx'
